Export Grafana dashboards from stats.usegalaxy.eu
Export dashboards using a GitHub Actions workflow, which replaces `export.sh`.  The latter reads the dashboards from the database and is designed to run periodically as a cron job on the stats host. The former reads the dashboards from the Grafana API and runs periodically on GitHub actions.

The motivation behind exporting the dashboards via the Grafana API is that it will guarantee that private dashboards do not get to this Git repository, as the workflow has the same permissions as any anonymous user.